+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = "2022, Caio Castro"
 author = "Caio Castro"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.1-alpha"
+release = "0.0.1"
 
 
 # -- General configuration ---------------------------------------------------
@@ -58,3 +58,6 @@ html_theme = "sphinx_book_theme"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+
+# Other configurations
+pygments_style = 'sphinx'