+++ fuzzy_machines/engine.py
@@ -5,9 +5,8 @@ from typing import Any, Dict
 from black import Union
 
 from fuzzy_machines.operands import OperandEnum
-
-from .kernel import Kernel
-from .rules import RuleBase
+from fuzzy_machines.kernel import Kernel
+from fuzzy_machines.rules import RuleBase
 
 
 class Engine:
@@ -16,20 +15,20 @@ class Engine:
     fuzzyfy, defuzzyfy and generate surface methods.
 
     Usage steps:
-    1: create engine
-    2: add input kernels (see Kernel for further reference)
-    3: add inference system kernel (see Kernel for further reference)
-    4: add rules to map the input kernels to the inference system
-    5: call engine.fuzzyfy() to run the system
-    6: call engine.defuzzyfy() to reduce the fuzzy result to a single float number
-    7: call engine.gen_surface() to build a iterable cache-like map to greatly reduce time compute
+    1: create engine \n
+    2: add input kernels (see Kernel for further reference) \n
+    3: add inference system kernel (see Kernel for further reference) \n
+    4: add rules to map the input kernels to the inference system \n
+    5: call engine.fuzzyfy() to run the system \n
+    6: call engine.defuzzyfy() to reduce the fuzzy result to a single float number \n
+    7: call engine.gen_surface() to build a iterable cache-like map to greatly reduce time compute 
     """
 
     def __init__(self, operands: OperandEnum = OperandEnum.DEFAULT) -> None:
         """Initializes a new engine object
 
         Args:
-            operands (OperandEnum, optional): Operation definitions for AND, OR and NOT methods.
+            operands (OperandEnum, optional): Operation definitions for AND, OR and NOT methods. \
         Defaults to OperandEnum.DEFAULT.
         """
         # initialization
@@ -58,12 +57,12 @@ class Engine:
         functions
 
         Args:
-            name (str): The name of the kernel. Various methods will call the kernel using this
+            name (str): The name of the kernel. Various methods will call the kernel using this \
         name as the key to a dict of type {name: kernel}
             kernel (Kernel): Kernel object, mapping the variable to its many membership functions
 
         Raises:
-            TypeError: if the internal dictionary self.input_kernel_set gets corrupted by direct
+            TypeError: if the internal dictionary self.input_kernel_set gets corrupted by direct \
         user manipulation
 
         Returns:
@@ -120,15 +119,15 @@ class Engine:
         system membership functions.
 
         Args:
-            name (str): must match the name of a KernelFuncMember registered at the Inference
+            name (str): must match the name of a KernelFuncMember registered at the Inference \
         Kernel System.
-            rule (Union[RuleBase, Dict[str, str]]): a rule object (AND, OR, NOT), or a dictionary
+            rule (Union[RuleBase, Dict[str, str]]): a rule object (AND, OR, NOT), or a dictionary \
         to get the direct value of a specific membership function.
 
         Examples:
-            - (i) If FOOD is GOOD then tip is High (direct access):
+            - (i) If FOOD is GOOD then tip is High (direct access): \
         fm.add_rule('High', {'food': 'good'})
-            - (ii) if SERVICE is BAD AND FOOD is RANCID then tip is Low (rule base access):
+            - (ii) if SERVICE is BAD AND FOOD is RANCID then tip is Low (rule base access): \
         fm.add_rule('Low', AND({'service': 'bad'}, {'food': 'rancid'}))
 
         See More:
@@ -158,17 +157,17 @@ class Engine:
 
     def fuzzyfy(self, measurements: Dict[str, Any]) -> Dict[str, float]:
         """
-        Main fuzzyfication method. Passing a dictionary of measurements for each registered
-        input kernel (in the format {'kernel_name': data}), this methods runs all kernel membership
-        functions and store the corresponding values.
+        Main fuzzyfication method. Passing a dictionary of measurements for each registered \
+        input kernel (in the format {'kernel_name': data}), this methods runs all kernel membership \
+        functions and store the corresponding values. \n
 
-        Next, it parses all rules and construct the "actuation signal", which is the actual values
-        sent to the inference system.
+        Next, it parses all rules and construct the "actuation signal", which is the actual values \
+        sent to the inference system. \n
 
-        Finally, the inference system is run based on the actuation signal.
+        Finally, the inference system is run based on the actuation signal. \n
 
         Args:
-            measurements (Dict[str, Any]): measurements for each registered registered input kernel
+            measurements (Dict[str, Any]): measurements for each registered registered input kernel \
         (in the format {'kernel_name': data})
 
         Raises: