+++ docs/source/HISTORY.rst
@@ -1,8 +1,12 @@
 =======
 History
 =======
-0.0.1 (2020–05–15)
+0.0.1 (2022–01–18)
 — — — — — — — — -
 
 * First release on PyPI.
+* Created Engine, Kernel, Rules, Operand and Rules modules
+* Library builds the Engine and map data input to fuzzy output, according to Kernel, Rules and Operands
+* TODO: missing an Engine.defuzzyfy method.
+* TODO: missing an Engine.generate surface method, for caching full set calculations